# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bondhon Matrimony API",
        "description": "Full-stack matrimony platform API. All protected endpoints require Bearer token from Sanctum.",
        "contact": {
            "email": "admin@bondhon.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/photos/pending": {
            "get": {
                "tags": [
                    "Admin - Photo Moderation"
                ],
                "summary": "Get photos pending moderation",
                "operationId": "f7a420b82eb9105ee4e26e5ff603987f",
                "responses": {
                    "200": {
                        "description": "Pending photos list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/photos/{id}/approve": {
            "put": {
                "tags": [
                    "Admin - Photo Moderation"
                ],
                "summary": "Approve a photo",
                "operationId": "dd98babdcc490e760a381f666c39fae7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Photo approved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/photos/{id}/reject": {
            "put": {
                "tags": [
                    "Admin - Photo Moderation"
                ],
                "summary": "Reject a photo",
                "operationId": "6a7fb0b74941408144d4b6a16d2925a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Inappropriate content"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Photo rejected"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/users": {
            "get": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "List all users with pagination",
                "operationId": "b0b00cf5db5258ff35b911db295a30da",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/users/{id}/ban": {
            "put": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Ban or unban a user",
                "operationId": "83fed388a4d5d92865f0e0b255709240",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "is_banned": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Violation of terms"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User ban status updated"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/users/{id}/verify": {
            "put": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Verify a user profile",
                "operationId": "d84cf165f181bb7bf226c472cd23bcc4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User verified"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token. Obtain via POST /api/v1/auth/login",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}